+++ backend/app/apis/super_admin_api.py
@@ -1,40 +1,9 @@
-from flask import jsonify, request, current_app
+from flask import request
+from ..models.data_cleaning import *
 from ..models.super_admin_model import SuperAdmin
-from datetime import datetime
+from .api_logger import *
 
 superadminModel = SuperAdmin()
-
-# function to log and return result
-# note that all result that is an error should be a string containing the word "Error" at the start
-# for ipAddress, if ngix is used for production, request.environ.get('HTTP_X_REAL_IP', request.remote_addr) should be used to get ip
-# actionDescription will be used for logging purposes
-
-def return_result(ipAddress, actionDescription, functionCalled, result):
-    
-    print("----Log Test----")
-
-    # who
-    print("Who: {}".format(ipAddress))
-
-    #what
-    print("What: {}".format(actionDescription))
-
-    #where
-    print("Where: {}".format(functionCalled))
-
-    #when
-    now = datetime.now()
-    print("When: {}".format(now))
-
-    #result
-    print("Result: {}".format(result))
-    print("----------------")
-    # checks if the result is an error result
-    if type(result) == str and "Error" in result:
-        return(jsonify(result), 401)
-    else:
-        return(jsonify(result), 201)
-
 
 def create_admin_account():
     if request.method == "POST":
@@ -44,24 +13,40 @@ def create_admin_account():
         password = request.form.get("Password")
         contactNumber = request.form.get("ContactNumber")
         
-        # if logged in user is a superadmin, create admin
-        if superadminModel.validate_super_admin(superAdminEmail):
-            return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Creating admin account", "create_admin_account", superadminModel.create_admin_account(username, adminEmail, password, contactNumber)))
+        # checks if the values are not none
+        if superAdminEmail is not None and username is not None and adminEmail is not None and password is not None and contactNumber is not None:
+
+            # checks if values are valid, if it is, proceed to create admin
+            if isemail(superAdminEmail) and isemail(superAdminEmail) and isstring(username) and isvalidpassword(password) and isvalidcontactnumber(contactNumber):
         
+                # if logged in user is a superadmin, create admin
+                if superadminModel.validate_super_admin(superAdminEmail):
+                    return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Creating admin account", "create_admin_account", superadminModel.create_admin_account(username, adminEmail, password, contactNumber)))
+                
+                else:
+                    return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Creating admin account", "create_admin_account", "Error User does not have permission"))
+            
+            else:
+                return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Creating admin account", "create_admin_account", "Error invalid input found"))               
+
         else:
-            return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Creating admin account", "create_admin_account", "Error User does not have permission"))
+            return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Creating admin account", "create_admin_account", "Error required values are empty"))    
 
 def search_admin():
     if request.method == "POST":
         adminEmail = request.form.get("AdminEmail")
         superAdminEmail = request.form.get("SuperAdminEmail")
 
-        # checks if the email calling the api is a Admin email, if it is, call the required functions
-        if superadminModel.validate_super_admin(superAdminEmail):
-            return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "SuperAdmin searching admin", "search_admin", superadminModel.search_admin(adminEmail)))
+        # checks if the values are valid and not none, if valid and not none check if the user has right role
+        if adminEmail is not None and isemail(adminEmail) and superAdminEmail is not None and isemail(superAdminEmail):
+            # checks if the email calling the api is a Admin email, if it is, call the required functions
+            if superadminModel.validate_super_admin(superAdminEmail):
+                return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "SuperAdmin searching admin", "search_admin", superadminModel.search_admin(adminEmail)))
+            else:
+                return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "SuperAdmin searching admin", "search_admin", "Error User does not have permission"))
+        
         else:
-            return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "SuperAdmin searching admin", "search_admin", "Error User does not have permission"))
-
+            return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "SuperAdmin searching admin", "search_admin", "Error invalid input found"))
 
 
 def delete_admin_account():
@@ -70,20 +55,30 @@ def delete_admin_account():
             superAdminEmail = request.form.get("superAdminEmail")
             adminEmail = request.form.get("AdminEmail")
 
-            # if logged in user is a superadmin, create admin
-            if superadminModel.validate_super_admin(superAdminEmail):
-                return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Deleting admin account", "delete_admin_account", superadminModel.delete_admin_account(adminEmail)))
-        
+            # checks if the values are valid and not none, if valid and not none check if the user has right role
+            if adminEmail is not None and isemail(adminEmail) and superAdminEmail is not None and isemail(superAdminEmail):
+                # if logged in user is a superadmin, create admin
+                if superadminModel.validate_super_admin(superAdminEmail):
+                    return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Deleting admin account", "delete_admin_account", superadminModel.delete_admin_account(adminEmail)))
+            
+                else:
+                    return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Deleting admin account", "delete_admin_account", "Error User does not have permission"))
             else:
-                return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Deleting admin account", "delete_admin_account", "Error User does not have permission"))
+                return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Deleting admin account", "delete_admin_account", "Error invalid input found"))
 
 def disable_admin_account():
     if request.method == "POST":
         superAdminEmail = request.form.get("superAdminEmail")
-        admninEmail = request.form.get("AdminEmail")
+        adminEmail = request.form.get("AdminEmail")
 
-        # if logged in user is a superadmin, disable admin account
-        if superadminModel.validate_super_admin(superAdminEmail):
-            return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Disabling admin account", "disable_admin_account", superadminModel.disable_admin_account(admninEmail)))
+
+        # checks if the values are valid and not none, if valid and not none check if the user has right role
+        if adminEmail is not None and isemail(adminEmail) and superAdminEmail is not None and isemail(superAdminEmail):
+            # if logged in user is a superadmin, disable admin account
+            if superadminModel.validate_super_admin(superAdminEmail):
+                return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Disabling admin account", "disable_admin_account", superadminModel.disable_admin_account(adminEmail)))
+            else:
+                return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Disabling admin account", "disable_admin_account", "Error User does not have permission"))
+
         else:
-            return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Disabling admin account", "disable_admin_account", "Error User does not have permission"))
+            return(return_result(request.environ.get('HTTP_X_REAL_IP', request.remote_addr), "Disabling admin account", "disable_admin_account", "Error invalid input found"))
